Fix duplicate guid for rerun of episode 40
The Spotify support team says this duplicate guid is causing their
system to fail to update, such that it does not include any new
episodes. This is strange, since this guid hasn't changed in over a
year, but I'm making this change in the hopes that it will fix the issue
for Spotify.