Fix broken audio player on episode details page
The audio_player.html was expectig a `post` variable, but this page was
passing in a `page` variable. Now, the audio_player.html expects two
parameters: playback_url and playback_guid. Anyone who includes it must
specify these two parameters.

This commit updates the two include sites to do this.


Former-commit-id: 2eff7eba9dc8f3de86ae0a7aa0fea0f8e7c44cca