Update the test feed to match the real feed
So I can start testing feed migration away from Feedburner.
Presently, the test feed is >1MB, so feedburner rejects it.

The next step is to create a new feedburner feed to this test feed
(https://softskills.audio/feed-test.xml), and then delete the test
feed from Feedburner, with the redirect option, and see if the
redirect option works.